+++ packages/in-memory/types/types.d.ts
@@ -1,6 +1,6 @@
-import { HistoryConstructor, HistoryOptions, History, ToArgument, LocationComponents, SessionLocation, PartialLocation, Hrefable } from "@hickory/root";
-export { HistoryConstructor, HistoryOptions, History, SessionLocation, PartialLocation, ToArgument, Hrefable, LocationComponents };
-export declare type InputLocations = Array<ToArgument>;
+import { HistoryConstructor, HistoryOptions, History, URLWithState, LocationComponents, SessionLocation, PartialLocation, Hrefable } from "@hickory/root";
+export { HistoryConstructor, HistoryOptions, History, SessionLocation, PartialLocation, URLWithState, Hrefable, LocationComponents };
+export declare type InputLocations = Array<URLWithState>;
 export interface SessionOptions {
     locations?: InputLocations;
     index?: number;
@@ -10,5 +10,5 @@ export interface InMemoryHistory extends History {
     reset(options?: SessionOptions): void;
 }
 export interface LocationOptions {
-    location: ToArgument;
+    location: URLWithState;
 }