+++ packages/location-utils/tests/location-utils.spec.ts
@@ -3,8 +3,7 @@ import {
   completePathname,
   completeHash,
   completeQuery,
-  stripPrefix,
-  stripBase
+  stripPrefix
 } from "../src";
 
 describe("location utils", () => {
@@ -66,34 +65,6 @@ describe("location utils", () => {
 
     it("does nothing when the string does not begin with the prefix", () => {
       expect(stripPrefix("/string", "/test")).toBe("/string");
-    });
-  });
-
-  describe("stripBase", () => {
-    it("strips the base off of the beginning of the path", () => {
-      const withBase = "/prefix/this-is-a-test";
-      expect(stripBase(withBase, "/prefix")).toBe("/this-is-a-test");
-    });
-
-    it("is not case-sensitive", () => {
-      const withBase = "/prefix/this-is-a-test";
-      expect(stripBase(withBase, "/PREFIX")).toBe("/this-is-a-test");
-    });
-
-    it("works when the first character after the base is a ?,#, or end of string", () => {
-      expect(stripBase("/prefix?query=true", "/prefix")).toBe("?query=true");
-      expect(stripBase("/prefix#yo", "/prefix")).toBe("#yo");
-      expect(stripBase("/prefix", "/prefix")).toBe("");
-    });
-
-    it("does nothing if the path does not contain the base segment", () => {
-      const noBase = "/this-is-a-test";
-      expect(stripBase(noBase, "/prefix")).toBe(noBase);
-    });
-
-    it("does not strip for incomplete prefix matches", () => {
-      const incompletePath = "/prefixthis-is-the-path";
-      expect(stripBase(incompletePath, "/prefix")).toBe(incompletePath);
     });
   });
 });