+++ packages/root/tests/locationUtils.spec.ts
@@ -1,22 +1,11 @@
 import "jest";
-import { locationUtils } from "../src";
+import { locationUtils, createBase } from "../src";
 import * as qs from "qs";
 
-import { SessionLocation, Key } from "../src/types";
+import { LocationComponents, Key } from "../src/types";
 
 describe("locationFactory", () => {
   describe("constructor", () => {
-    it("throws when attempting to use an invalid base", () => {
-      const badValues = ["does-not-start-with-a-slash", "/ends-with-slash/"];
-      badValues.forEach(value => {
-        expect(() => {
-          const creators = locationUtils({
-            base: value
-          });
-        }).toThrow();
-      });
-    });
-
     describe("query option", () => {
       const consoleWarn = console.warn;
 
@@ -71,13 +60,46 @@ describe("locationFactory", () => {
     describe("pathname", () => {
       describe("base", () => {
         const { location } = locationUtils({
-          base: "/prefix"
+          base: createBase("/prefix")
         });
 
         it("strips the base off of the string", () => {
           const loc = location({ url: "/prefix/this/is/the/rest" });
           expect(loc.pathname).toBe("/this/is/the/rest");
         });
+
+        it('sets pathname to "/" if pathname exactly equals base', () => {
+          const loc = location({ url: "/prefix" });
+          expect(loc.pathname).toBe("/");
+        });
+
+        it("throws if URL begins with a pathname but no base and base is strict", () => {
+          const { location } = locationUtils({
+            base: createBase("/prefix", { strict: true })
+          });
+          expect(() => {
+            const loc = location({ url: "/this/is/the/rest" });
+          }).toThrow(
+            'Expected a string that begins with "/prefix", but received "/this/is/the/rest".'
+          );
+        });
+
+        it("does not throw if URL begins with a query", () => {
+          expect(() => {
+            const loc = location({ url: "?test=ing" });
+          }).not.toThrow();
+        });
+
+        it("does not throw if URL begins with a hash", () => {
+          expect(() => {
+            const loc = location({ url: "#test" });
+          }).not.toThrow();
+        });
+      });
+
+      it("is parsed from the url string", () => {
+        const loc = location({ url: "/pathname?query=this#hash" });
+        expect(loc.pathname).toBe("/pathname");
       });
 
       describe("other parts", () => {
@@ -91,63 +113,17 @@ describe("locationFactory", () => {
           expect(output.pathname).toBe("/Chance");
         });
       });
-
-      describe("url object argument", () => {
-        it("is parsed from the url string", () => {
-          const loc = location({ url: "/pathname?query=this#hash" });
-          expect(loc.pathname).toBe("/pathname");
-        });
-      });
-
-      describe("partial object argument", () => {
-        it("uses provided pathname", () => {
-          const input = {
-            pathname: "/test",
-            query: "one=two",
-            hash: "hello"
-          };
-          const output = location(input);
-          expect(output.pathname).toBe("/test");
-        });
-
-        it("is / if no pathname is provided", () => {
-          const input = {
-            query: "one=two",
-            hash: "hello"
-          };
-          const output = location(input);
-          expect(output.pathname).toBe("/");
-        });
-      });
     });
 
     describe("query", () => {
-      describe("url object argument", () => {
-        it("is parsed from the url string", () => {
-          const loc = location({ url: "/pathname?query=this#hash" });
-          expect(loc.query).toBe("query=this");
-        });
+      it("is parsed from the url string", () => {
+        const loc = location({ url: "/pathname?query=this" });
+        expect(loc.query).toBe("query=this");
       });
 
-      describe("partial object argument", () => {
-        it("uses provided query", () => {
-          const input = {
-            pathname: "/test",
-            query: "one=two",
-            hash: "hello"
-          };
-          const output = location(input);
-          expect(output.query).toBe("one=two");
-        });
-
-        it("sets default value if none is provided", () => {
-          const input = {
-            pathname: "/test",
-            hash: "hello"
-          };
-          const output = location(input);
-          expect(output.query).toBe("");
-        });
+      it("does not include a hash", () => {
+        const loc = location({ url: "/pathname?query=this#hash" });
+        expect(loc.query).toBe("query=this");
       });
 
       describe("query.parse option", () => {
@@ -165,62 +141,114 @@ describe("locationFactory", () => {
     });
 
     describe("hash", () => {
-      describe("url object argument", () => {
-        it("is parsed from the url string", () => {
-          const loc = location({ url: "/pathname?query=this#hash" });
-          expect(loc.hash).toBe("hash");
-        });
-      });
-
-      describe("partial object argument", () => {
-        it("uses provided hash", () => {
-          const input = {
-            pathname: "/test",
-            query: "one=two",
-            hash: "hello"
-          };
-          const output = location(input);
-          expect(output.hash).toBe("hello");
-        });
-
-        it("sets hash to empty string if none is provided", () => {
-          const input = {
-            pathname: "/test",
-            query: "one=two"
-          };
-          const output = location(input);
-          expect(output.hash).toBe("");
-        });
+      it("is parsed from the url string", () => {
+        const loc = location({ url: "/pathname?query=this#hash" });
+        expect(loc.hash).toBe("hash");
       });
     });
 
     describe("state", () => {
-      describe("url object argument", () => {
-        it("is the state property from the object", () => {
-          const state = {
-            omg: "bff"
-          };
-          const input = {
-            url: "/",
-            state
-          };
-          const output = location(input);
-          expect(output.state).toBeDefined();
-          expect(output.state).toMatchObject(state);
-        });
-      });
-
-      describe("partial object argument", () => {
-        it("adds state if provided", () => {
-          const state = { fromLocation: false };
-          const output = location({ pathname: "/", state });
-          expect(output.state).toEqual(state);
-        });
+      it("is the state property from the object", () => {
+        const state = {
+          omg: "bff"
+        };
+        const input = {
+          url: "/",
+          state
+        };
+        const output = location(input);
+        expect(output.state).toBeDefined();
+        expect(output.state).toMatchObject(state);
       });
 
       it("is undefined if not provided", () => {
         const output = location({ url: "/" });
         expect(output.state).toBeUndefined();
+      });
+    });
+
+    describe("special cases", () => {
+      describe("hash fragment URLs", () => {
+        describe("with current", () => {
+          const current: LocationComponents = {
+            pathname: "/test",
+            query: "one=two",
+            hash: "hooray"
+          };
+
+          it("re-uses current pathname", () => {
+            const loc = location({ url: "#hash" }, current);
+            expect(loc.pathname).toBe("/test");
+          });
+
+          it("re-uses current query", () => {
+            const loc = location({ url: "#hash" }, current);
+            expect(loc.query).toBe("one=two");
+          });
+
+          it("uses provided hash", () => {
+            const loc = location({ url: "#hash" }, current);
+            expect(loc.hash).toBe("hash");
+          });
+        });
+
+        describe("without current", () => {
+          it('pathname is "/"', () => {
+            const loc = location({ url: "#hash" });
+            expect(loc.pathname).toBe("/");
+          });
+
+          it("query is an empty value", () => {
+            const loc = location({ url: "#hash" });
+            expect(loc.query).toBe("");
+          });
+
+          it("uses provided hash", () => {
+            const loc = location({ url: "#hash" });
+            expect(loc.hash).toBe("hash");
+          });
+        });
+      });
+
+      describe("empty string", () => {
+        describe("with current", () => {
+          const current: LocationComponents = {
+            pathname: "/test",
+            query: "one=two",
+            hash: "hey"
+          };
+          it("re-uses current pathname", () => {
+            const loc = location({ url: "" }, current);
+            expect(loc.pathname).toBe("/test");
+          });
+
+          it("re-uses current query", () => {
+            const loc = location({ url: "" }, current);
+            expect(loc.query).toBe("one=two");
+          });
+
+          it("re-uses current hash", () => {
+            const loc = location({ url: "" }, current);
+            expect(loc.hash).toBe("hey");
+          });
+        });
+
+        describe("without current", () => {
+          it('pathname is "/"', () => {
+            const loc = location({ url: "" });
+            expect(loc.pathname).toBe("/");
+          });
+
+          it("query is empty value", () => {
+            const loc = location({ url: "" });
+            expect(loc.query).toBe("");
+          });
+
+          it("hash is empty string", () => {
+            const loc = location({ url: "" });
+            expect(loc.hash).toBe("");
+          });
+        });
       });
     });
   });
@@ -265,7 +293,7 @@ describe("locationFactory", () => {
       describe("base", () => {
         it("adds the base to the generated string", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const loc = {
             pathname: "/one/two/three",
@@ -278,7 +306,7 @@ describe("locationFactory", () => {
 
         it("does not include the base if there is no pathname", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const loc = {
             query: "?test=ing",
@@ -385,7 +413,7 @@ describe("locationFactory", () => {
 
         it("prefixes with base", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const path = stringify("/one/two/three#four");
           expect(path).toBe("/prefix/one/two/three#four");
@@ -393,7 +421,7 @@ describe("locationFactory", () => {
 
         it("prefixes pathname when joining with base", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const path = stringify("one");
           expect(path).toBe("/prefix/one");
@@ -408,7 +436,7 @@ describe("locationFactory", () => {
 
         it("if there is a base, it is not prepended", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const path = stringify("?test=true");
           expect(path).toBe("?test=true");
@@ -423,7 +451,7 @@ describe("locationFactory", () => {
 
         it("if there is a base, it is not prepended", () => {
           const { stringify } = locationUtils({
-            base: "/prefix"
+            base: createBase("/prefix")
           });
           const path = stringify("#test");
           expect(path).toBe("#test");