+++ packages/root/types/base.d.ts
@@ -1,0 +1,2 @@
+import { BaseFunctions, BaseOptions } from "./types/base";
+export default function createBase(base: string, options?: BaseOptions): BaseFunctions;