+++ packages/root/types/index.d.ts
@@ -3,4 +3,5 @@ import locationUtils from "./locationUtils";
 import keyGenerator from "./keyGenerator";
 import navigateWith from "./navigateWith";
 import navigationConfirmation from "./navigationConfirmation";
-export { locationUtils, keyGenerator, navigationConfirmation, navigateWith };
+import createBase from "./base";
+export { locationUtils, keyGenerator, navigationConfirmation, navigateWith, createBase };