+++ packages/root/types/types/base.d.ts
@@ -1,0 +1,8 @@
+export interface BaseFunctions {
+    add(pathname: string): string;
+    remove(pathname: string): string;
+}
+export interface BaseOptions {
+    emptyRoot?: boolean;
+    strict?: boolean;
+}