+++ packages/root/types/types/index.d.ts
@@ -1,6 +1,6 @@
 export { HistoryOptions, HistoryConstructor, History } from "./hickory";
-export { LocationComponents, PartialLocation, SessionLocation, Hrefable, Key } from "./location";
+export { LocationComponents, PartialLocation, SessionLocation, Hrefable, Key, URLWithState } from "./location";
 export { KeyFns } from "./keyGenerator";
 export { LocationUtilOptions, QueryFunctions, LocationUtils } from "./locationUtils";
 export { NavigationInfo, ConfirmationFunction, ConfirmationMethods, BlockingHistory } from "./navigationConfirmation";
-export { ToArgument, Action, NavType, FinishNavigation, CancelNavigation, PendingNavigation, ResponseHandler, Preparer, FinishCancel, NavigateArgs, NavigateHelpers } from "./navigate";
+export { Action, NavType, FinishNavigation, CancelNavigation, PendingNavigation, ResponseHandler, Preparer, FinishCancel, NavigateArgs, NavigateHelpers } from "./navigate";