+++ packages/root/types/types/locationUtils.d.ts
@@ -1,14 +1,15 @@
-import { LocationComponents, SessionLocation, Hrefable, Key, PartialLocation, URLWithState } from "./location";
+import { LocationComponents, SessionLocation, Hrefable, Key, URLWithState } from "./location";
+import { BaseFunctions } from "./base";
 export interface QueryFunctions {
-    parse: (query?: string) => any;
-    stringify: (query?: any) => string;
+    parse(query?: string): any;
+    stringify(query?: any): string;
 }
 export interface LocationUtilOptions {
     query?: QueryFunctions;
-    base?: string;
+    base?: BaseFunctions;
 }
 export interface LocationUtils {
     keyed(location: LocationComponents, key: Key): SessionLocation;
-    location(value: PartialLocation | URLWithState): LocationComponents;
+    location(value: URLWithState, current?: LocationComponents): LocationComponents;
     stringify(location: Hrefable): string;
 }